requirements.txt
discord.py>=2.4,<3.0
aiosqlite>=0.20
python-dotenv>=1.0
flask>=3.0
